+++ tests/unit/neighbor_list/test_neighbor_list_nvalchemi.py
@@ -97,6 +97,7 @@ def test_neighborlist_nvalchemi(name, frame, cutoff, self_interaction):
     for met_name in method_list:
         dds = []
         for data in dataloader:
+            data.cell = data.cell.to(data.pos.dtype)
             if met_name == "ase_ref":
                 met = ase_neighbor_list
             else:
@@ -149,6 +150,7 @@ def test_neighborlist_pbc_nvalchemi(nls_name):
                 # Get nvalchemi neighbor list distances
                 nvalchemi_distances = []
                 for data in dataloader:
+                    data.cell = data.cell.to(data.pos.dtype)
                     if "cell" in data:
                         print("Cell:\n", data.cell)
                     idx_i, idx_j, cell_shifts, _ = nls_method(
@@ -171,6 +173,7 @@ def test_neighborlist_pbc_nvalchemi(nls_name):
                 # Get ASE reference distances
                 ase_distances = []
                 for data in dataloader:
+                    data.cell = data.cell.to(data.pos.dtype)
                     idx_i, idx_j, ase_cell_shifts, _ = ase_neighbor_list(
                         data, cutoff, self_interaction=self_interaction
                     )
@@ -212,6 +215,7 @@ def test_pbc_minimum_image_convention_nvalchemi(nls_name):
     dataloader = DataLoader(data_list, batch_size=1)
 
     for data in dataloader:
+        data.cell = data.cell.to(data.pos.dtype)
         idx_i, idx_j, cell_shifts, _ = nls_method(
             data, cutoff, self_interaction=False
         )
@@ -230,13 +234,14 @@ def test_pbc_minimum_image_convention_nvalchemi(nls_name):
 
     distances = []
     for data in dataloader:
+        data.cell = data.cell.to(data.pos.dtype)
         idx_i, idx_j, cell_shifts, _ = nls_method(
             data, cutoff, self_interaction=False
         )
         if nls_name == "raw":
             cell = data.cell.reshape(-1, 3, 3)
             cell_shifts = (
-                cell_shifts.to(cell.dtype).to(cell.dtype).unsqueeze(-1)
+                cell_shifts.to(cell.dtype).unsqueeze(-1)
                 * cell[data.batch[idx_i]]
             ).sum(dim=1)
 
@@ -277,6 +282,7 @@ def test_mixed_pbc_nvalchemi(nls_name):
     dataloader = DataLoader(data_list, batch_size=1)
     distances = []
     for data in dataloader:
+        data.cell = data.cell.to(data.pos.dtype)
         idx_i, idx_j, cell_shifts, _ = nls_method(
             data, cutoff, self_interaction=False
         )
